+++ src/backend/src/main/java/com/personal/finance/backend/reports/dto/response/TrendDataDTO.java
@@ -5,12 +5,14 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
 public class TrendDataDTO {
     private String date;
-    private Double income;
-    private Double expense;
+    private BigDecimal income;
+    private BigDecimal expense;
 }
